+++ stream-chat-android-client/build.gradle
@@ -51,6 +51,7 @@ tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
         freeCompilerArgs += [
                 '-progressive',
                 '-Xexplicit-api=strict',
+                '-Xopt-in=kotlin.RequiresOptIn',
                 '-Xopt-in=io.getstream.chat.android.core.internal.InternalStreamChatApi',
         ]
     }
