+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/experimental/plugin/Plugin.kt
@@ -1,13 +1,15 @@
-package io.getstream.chat.android.client.plugin
+package io.getstream.chat.android.client.experimental.plugin
 
 import android.content.Context
 import io.getstream.chat.android.client.ChatClient
-import io.getstream.chat.android.client.plugin.listeners.OperationListenersFacade
+import io.getstream.chat.android.client.experimental.plugin.listeners.OperationListenersFacade
+import io.getstream.chat.android.core.ExperimentalStreamChatApi
 
 /**
  * Plugin is an extension for [ChatClient]. It extends [OperationListenersFacade] and provides additionally
  * name attribute and init method.
  */
+@ExperimentalStreamChatApi
 public interface Plugin : OperationListenersFacade {
     public val name: String
