+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/experimental/plugin/listeners/OperationListenersFacade.kt
@@ -1,0 +1,6 @@
+package io.getstream.chat.android.client.experimental.plugin.listeners
+
+import io.getstream.chat.android.core.ExperimentalStreamChatApi
+
+@ExperimentalStreamChatApi
+public interface OperationListenersFacade : QueryChannelsListener, SendMessageListener
