+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/experimental/plugin/listeners/QueryChannelsListener.kt
@@ -1,10 +1,12 @@
-package io.getstream.chat.android.client.plugin.listeners
+package io.getstream.chat.android.client.experimental.plugin.listeners
 
 import io.getstream.chat.android.client.api.models.QueryChannelRequest
 import io.getstream.chat.android.client.api.models.QueryChannelsRequest
 import io.getstream.chat.android.client.models.Channel
 import io.getstream.chat.android.client.utils.Result
+import io.getstream.chat.android.core.ExperimentalStreamChatApi
 
+@ExperimentalStreamChatApi
 public interface QueryChannelsListener {
     public suspend fun onQueryChannelRequest(
         channelType: String,
