+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/experimental/plugin/listeners/SendMessageListener.kt
@@ -1,8 +1,10 @@
-package io.getstream.chat.android.client.plugin.listeners
+package io.getstream.chat.android.client.experimental.plugin.listeners
 
 import io.getstream.chat.android.client.models.Message
 import io.getstream.chat.android.client.utils.Result
+import io.getstream.chat.android.core.ExperimentalStreamChatApi
 
+@ExperimentalStreamChatApi
 public interface SendMessageListener {
     public fun onMessageSendRequest(
         channelType: String,
