--- stream-chat-android-client/src/main/java/io/getstream/chat/android/client/plugin/listeners/OperationListenersFacade.kt
@@ -1,3 +1,0 @@
-package io.getstream.chat.android.client.plugin.listeners
-
-public interface OperationListenersFacade : QueryChannelsListener, SendMessageListener
