+++ stream-chat-android-offline/api/stream-chat-android-offline.api
@@ -410,14 +410,37 @@ public final class io/getstream/chat/android/offline/channel/ChannelData {
 	public fun toString ()Ljava/lang/String;
 }
 
-public final class io/getstream/chat/android/offline/extensions/ChatClientKt {
+public final class io/getstream/chat/android/offline/experimental/extensions/ChatClientKt {
+}
+
+public final class io/getstream/chat/android/offline/experimental/plugin/OfflinePlugin$Companion {
+}
+
+public final class io/getstream/chat/android/offline/experimental/querychannels/state/ChannelsState$Loading : io/getstream/chat/android/offline/experimental/querychannels/state/ChannelsState {
+	public static final field INSTANCE Lio/getstream/chat/android/offline/experimental/querychannels/state/ChannelsState$Loading;
+}
+
+public final class io/getstream/chat/android/offline/experimental/querychannels/state/ChannelsState$NoQueryActive : io/getstream/chat/android/offline/experimental/querychannels/state/ChannelsState {
+	public static final field INSTANCE Lio/getstream/chat/android/offline/experimental/querychannels/state/ChannelsState$NoQueryActive;
+}
+
+public final class io/getstream/chat/android/offline/experimental/querychannels/state/ChannelsState$OfflineNoResults : io/getstream/chat/android/offline/experimental/querychannels/state/ChannelsState {
+	public static final field INSTANCE Lio/getstream/chat/android/offline/experimental/querychannels/state/ChannelsState$OfflineNoResults;
+}
+
+public final class io/getstream/chat/android/offline/experimental/querychannels/state/ChannelsState$Result : io/getstream/chat/android/offline/experimental/querychannels/state/ChannelsState {
+	public fun <init> (Ljava/util/List;)V
+	public final fun component1 ()Ljava/util/List;
+	public final fun copy (Ljava/util/List;)Lio/getstream/chat/android/offline/experimental/querychannels/state/ChannelsState$Result;
+	public static synthetic fun copy$default (Lio/getstream/chat/android/offline/experimental/querychannels/state/ChannelsState$Result;Ljava/util/List;ILjava/lang/Object;)Lio/getstream/chat/android/offline/experimental/querychannels/state/ChannelsState$Result;
+	public fun equals (Ljava/lang/Object;)Z
+	public final fun getChannels ()Ljava/util/List;
+	public fun hashCode ()I
+	public fun toString ()Ljava/lang/String;
 }
 
 public final class io/getstream/chat/android/offline/extensions/ChatErrorKt {
 	public static final fun isPermanent (Lio/getstream/chat/android/client/errors/ChatError;)Z
-}
-
-public final class io/getstream/chat/android/offline/plugin/OfflinePlugin$Companion {
 }
 
 public final class io/getstream/chat/android/offline/querychannels/QueryChannelsController {
@@ -456,32 +479,6 @@ public final class io/getstream/chat/android/offline/querychannels/QueryChannels
 	public final fun component1 ()Ljava/util/List;
 	public final fun copy (Ljava/util/List;)Lio/getstream/chat/android/offline/querychannels/QueryChannelsController$ChannelsState$Result;
 	public static synthetic fun copy$default (Lio/getstream/chat/android/offline/querychannels/QueryChannelsController$ChannelsState$Result;Ljava/util/List;ILjava/lang/Object;)Lio/getstream/chat/android/offline/querychannels/QueryChannelsController$ChannelsState$Result;
-	public fun equals (Ljava/lang/Object;)Z
-	public final fun getChannels ()Ljava/util/List;
-	public fun hashCode ()I
-	public fun toString ()Ljava/lang/String;
-}
-
-public abstract class io/getstream/chat/android/offline/querychannels/state/ChannelsState {
-}
-
-public final class io/getstream/chat/android/offline/querychannels/state/ChannelsState$Loading : io/getstream/chat/android/offline/querychannels/state/ChannelsState {
-	public static final field INSTANCE Lio/getstream/chat/android/offline/querychannels/state/ChannelsState$Loading;
-}
-
-public final class io/getstream/chat/android/offline/querychannels/state/ChannelsState$NoQueryActive : io/getstream/chat/android/offline/querychannels/state/ChannelsState {
-	public static final field INSTANCE Lio/getstream/chat/android/offline/querychannels/state/ChannelsState$NoQueryActive;
-}
-
-public final class io/getstream/chat/android/offline/querychannels/state/ChannelsState$OfflineNoResults : io/getstream/chat/android/offline/querychannels/state/ChannelsState {
-	public static final field INSTANCE Lio/getstream/chat/android/offline/querychannels/state/ChannelsState$OfflineNoResults;
-}
-
-public final class io/getstream/chat/android/offline/querychannels/state/ChannelsState$Result : io/getstream/chat/android/offline/querychannels/state/ChannelsState {
-	public fun <init> (Ljava/util/List;)V
-	public final fun component1 ()Ljava/util/List;
-	public final fun copy (Ljava/util/List;)Lio/getstream/chat/android/offline/querychannels/state/ChannelsState$Result;
-	public static synthetic fun copy$default (Lio/getstream/chat/android/offline/querychannels/state/ChannelsState$Result;Ljava/util/List;ILjava/lang/Object;)Lio/getstream/chat/android/offline/querychannels/state/ChannelsState$Result;
 	public fun equals (Ljava/lang/Object;)Z
 	public final fun getChannels ()Ljava/util/List;
 	public fun hashCode ()I
