+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/experimental/extensions/ChatClient.kt
@@ -1,12 +1,13 @@
-package io.getstream.chat.android.offline.extensions
+package io.getstream.chat.android.offline.experimental.extensions
 
 import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.core.ExperimentalStreamChatApi
 import io.getstream.chat.android.core.internal.InternalStreamChatApi
-import io.getstream.chat.android.offline.plugin.OfflinePlugin
-import io.getstream.chat.android.offline.plugin.adapter.ChatClientReferenceAdapter
-import io.getstream.chat.android.offline.plugin.state.StateRegistry
+import io.getstream.chat.android.offline.experimental.plugin.OfflinePlugin
+import io.getstream.chat.android.offline.experimental.plugin.adapter.ChatClientReferenceAdapter
+import io.getstream.chat.android.offline.experimental.plugin.state.StateRegistry
 
+@ExperimentalStreamChatApi
 internal val ChatClient.state: StateRegistry
     get() = requireNotNull((plugins.firstOrNull { it.name == OfflinePlugin.MODULE_NAME } as? OfflinePlugin)?.state) {
         "Offline plugin must be configured in ChatClient"
