+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/experimental/plugin/QueryReference.kt
@@ -1,4 +1,4 @@
-package io.getstream.chat.android.offline.plugin
+package io.getstream.chat.android.offline.experimental.plugin
 
 import io.getstream.chat.android.client.call.Call
 import io.getstream.chat.android.core.ExperimentalStreamChatApi
