+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/experimental/plugin/adapter/ChatClientReferenceAdapter.kt
@@ -1,10 +1,10 @@
-package io.getstream.chat.android.offline.plugin.adapter
+package io.getstream.chat.android.offline.experimental.plugin.adapter
 
 import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.client.api.models.QueryChannelsRequest
 import io.getstream.chat.android.core.ExperimentalStreamChatApi
 import io.getstream.chat.android.core.internal.InternalStreamChatApi
-import io.getstream.chat.android.offline.querychannels.QueryChannelsReference
+import io.getstream.chat.android.offline.experimental.querychannels.QueryChannelsReference
 
 @InternalStreamChatApi
 @ExperimentalStreamChatApi
