+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/experimental/plugin/state/StateRegistry.kt
@@ -1,4 +1,4 @@
-package io.getstream.chat.android.offline.plugin.state
+package io.getstream.chat.android.offline.experimental.plugin.state
 
 import io.getstream.chat.android.client.api.models.FilterObject
 import io.getstream.chat.android.client.api.models.QuerySort
@@ -7,8 +7,8 @@ import io.getstream.chat.android.core.ExperimentalStreamChatApi
 import io.getstream.chat.android.core.internal.InternalStreamChatApi
 import io.getstream.chat.android.offline.ChatDomain
 import io.getstream.chat.android.offline.ChatDomainImpl
-import io.getstream.chat.android.offline.querychannels.state.QueryChannelsMutableState
-import io.getstream.chat.android.offline.querychannels.state.QueryChannelsState
+import io.getstream.chat.android.offline.experimental.querychannels.state.QueryChannelsMutableState
+import io.getstream.chat.android.offline.experimental.querychannels.state.QueryChannelsState
 import kotlinx.coroutines.CoroutineScope
 import java.util.concurrent.ConcurrentHashMap
