+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/experimental/querychannels/QueryChannelsReference.kt
@@ -1,4 +1,4 @@
-package io.getstream.chat.android.offline.querychannels
+package io.getstream.chat.android.offline.experimental.querychannels
 
 import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.client.api.models.QueryChannelsRequest
@@ -7,9 +7,9 @@ import io.getstream.chat.android.client.call.await
 import io.getstream.chat.android.client.models.Channel
 import io.getstream.chat.android.core.ExperimentalStreamChatApi
 import io.getstream.chat.android.core.internal.InternalStreamChatApi
-import io.getstream.chat.android.offline.extensions.state
-import io.getstream.chat.android.offline.plugin.QueryReference
-import io.getstream.chat.android.offline.querychannels.state.QueryChannelsState
+import io.getstream.chat.android.offline.experimental.extensions.state
+import io.getstream.chat.android.offline.experimental.plugin.QueryReference
+import io.getstream.chat.android.offline.experimental.querychannels.state.QueryChannelsState
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
