+++ stream-chat-android-offline/src/test/java/io/getstream/chat/android/offline/IdGenerationDomainTest.kt
@@ -6,7 +6,7 @@ import com.google.common.truth.Truth
 import com.nhaarman.mockitokotlin2.mock
 import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.client.models.User
-import io.getstream.chat.android.offline.plugin.OfflinePlugin
+import io.getstream.chat.android.offline.experimental.plugin.OfflinePlugin
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
